+++ src/main/kotlin/com/comsysto/xmaslab2021/Location.kt
@@ -1,0 +1,6 @@
+package com.comsysto.xmaslab2021
+
+data class Location (
+    val altitude: Double,
+    val latitude: Double
+        )
