+++ src/main/kotlin/com/comsysto/xmaslab2021/offer/OfferController.kt
@@ -1,0 +1,18 @@
+package com.comsysto.xmaslab2021.offer
+
+import com.comsysto.xmaslab2021.offer.dto.OfferCreationModel
+import com.comsysto.xmaslab2021.offer.dto.OfferModel
+import org.springframework.web.bind.annotation.*
+
+@RestController
+@RequestMapping("/offers")
+class OfferController(private val offers: OfferService) {
+
+    @PutMapping("/{offerId}")
+    fun createOffer(@PathVariable offerId: String, @RequestBody offerCreation: OfferCreationModel){
+        offers.saveOffer(Offer(offerId, offerCreation.description, offerCreation.location))
+    }
+
+    @GetMapping
+    fun getAll(): List<OfferModel> = offers.allOffers().map { o -> OfferModel(o) }
+}
