+++ src/main/kotlin/com/comsysto/xmaslab2021/offer/OfferRepository.kt
@@ -1,0 +1,7 @@
+package com.comsysto.xmaslab2021.offer
+
+import org.springframework.data.jpa.repository.JpaRepository
+import org.springframework.stereotype.Repository
+
+@Repository
+interface OfferRepository : JpaRepository<Offer, String>
