+++ src/main/kotlin/com/comsysto/xmaslab2021/offer/OfferService.kt
@@ -1,0 +1,9 @@
+package com.comsysto.xmaslab2021.offer
+
+import org.springframework.stereotype.Service
+
+@Service
+class OfferService(private val offers: OfferRepository) {
+    fun allOffers(): List<Offer> = offers.findAll()
+    fun saveOffer(offer: Offer) = offers.save(offer)
+}
