+++ src/main/kotlin/com/comsysto/xmaslab2021/offer/dto/OfferCreationModel.kt
@@ -1,0 +1,5 @@
+package com.comsysto.xmaslab2021.offer.dto
+
+import com.comsysto.xmaslab2021.Location
+
+open class OfferCreationModel (val description: String, val location: Location)
